move bdiff to bin
drawing a line between what ends up as an alias and what ends up as a program. an alias points to a program and may contain flags. a program is any sequence of multiple commands. a program can be a single line of commands using pipes, etc, or a sequence of commands and other operations.